requirements.txt
streamlit
pandas
altair
supabase
python-dotenv